lib: rust_support: Correct signature of vmm_alloc_physical
Fix vmm_alloc_physical to more closly match the C version. It takes a
paddr_t argument for the physical address to map, not a pointer to a
paddr_t like vmm_alloc_physical_etc does (as that function needs an
array of physical addresses, not a single physical address).

Bug: 298705967
Change-Id: Ie4230a44c56d1896bdac6c0fa011b72ac9a435ca