Disable test storage for tests with run on sandbox property
Test storage is not accessible to SDK sandboxes since it is a content
provider that is not accessible to sandboxes. So far, test storage
access has been disabled for tests that use the instrument sandbox tag
directly. This change ensures that it is disabled for all kinds of
sandbox tests including ones defined using the run_on_sdk_sandbox
property.

Bug: 325423567
Test: m tradefed-all
Test: Run inprocess tests and tests with run_on_sdk_sandbox tag with
changes applied.

Change-Id: I92692ca15a88a3d020ccadb01f8a6b496a44d9d0