Prevent PeriodicProvisioner from running in e2e tests
Since the end to end tests run in the context of rkpdapp, they face
the risk that PeriodicProvisioner runs while the test also runs.

Prevent this by adding a serialization lock around the periodic job,
so that it cannot happen while tests run (and vice versa).

PeriodicProvisioner is always supposed to be serial anyway, so adding
the lock has no impact on normal execution of the background job.

Bug: 330756446
Test: RkpdAppIntegrationTests

Change-Id: I11ceee77c8681869d8b7eeff3e783180d48a1749