[scudo] Change tests that use setrlimit to cause mmap to fail. (#87004)
It appears that qemu does not actually cause mmap to fail when calling
setrlimit to limit the address space size. In the two tests that use
setrlimit, detect if mmap still works and skip the tests in that case.

Since all Android targets should support setrlimit, compile out the mmap
check code for them.

GitOrigin-RevId: c0a3c5c81ff7ff9e2a90b1664d6e419bf99c1aa9
Change-Id: If0779cb50a77f3fea8e6704d4a3d21c3eaa47e3b