Add debug info for failure to become 1-threaded
Add more information for the 'Failed to reach single-threaded state'
crashes in the zygote.

We capture abbreviated /proc/.../stat for all other threads
in the process at that point. After doing so, we also re-read and
re-print the line for the requesting process, to see if anything
changed while we were capturing the others.

This should tell us whether we're still dealing with the expected
daemons not exiting fast enough, or whether we have a different issue
as well.

This involved moving the /proc/.../stat printing code to enable reuse.

We also now use the same code in the primary test to determine if we
are single-threaded at the OS level.

Test: Standalone test program attached to b/310106506.
Test: art/test/testrunner/testrunner.py --host -b with kNumTries = 1
Bug: 310106506
Bug: 333219584
Bug: 315721088
Bug: 314247665
Bug: 314225311
Bug: 319209976
Bug: 317838599
Change-Id: I09aa9dd1bca13f7cd3bfcc1551bceaf1eb7458c2