scripts: Stop incorrectly reporting tests as flakes
Previously, `run_test` was using its `retry` arg (whether a test should
be retried on failure) to populate the `TestResult` `retried` field
(whether the test _actually was_ retried). This made the "might be
flaky" warning get printed for any test that passes the first time, but
not for tests which pass when retried.

Bug: None
Change-Id: I6602149067ad8242ec182788146713b0c854e1e3